+++ src/components/Footer.tsx
@@ -1,0 +1,13 @@
+import '../styles/footer.scss'
+
+const Footer = () => {
+    return(
+        <footer>
+            <small>
+                <span>&copy; 2025 Alfonsas Alubeckas. All Rights Reserved.</span>
+            </small>
+        </footer>
+    ) 
+  }
+  
+  export default Footer
